ai-native-dev-main 2/labs/06-observability/demo-05-rag-pipeline-observability/pyproject.toml
[project]
name = "demo-05-rag-pipeline-observability"
version = "0.1.0"
description = "End-to-end observability for a RAG pipeline with LangChain and LangSmith"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langsmith>=0.1.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-core>=0.3.0",
    "langchain-community>=0.3.0",
    "chromadb>=0.5.0",
    "langchain-chroma>=0.1.0",
    "python-dotenv>=1.0.0",
]
